# lib/api-zod/src/generated/types/authResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Clownin API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfile } from './userProfile';

export interface AuthResponse {
  token: string;
  user: UserProfile;
}