# lib/api-zod/src/generated/types/createFileBody.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Clownin API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateFileBody {
  /** @minLength 1 */
  path: string;
  content?: string;
  language?: string;
}