# lib/api-zod/src/generated/types/createProjectBody.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Clownin API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProjectBody {
  /** @minLength 1 */
  name: string;
  language?: string;
  description?: string;
}